Fixed: The import path error caused by moving the directory of the flight booking AI-Agent.
Fix import and remove `config.example.yml` file.
