+++ book-flight-ai-agent/go-client/cmd/client.go
@@ -31,8 +31,8 @@ import (
 )
 
 import (
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/conf"
-	chat "github.com/apache/dubbo-go-samples/llm/book-flight/proto"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/conf"
+	chat "github.com/apache/dubbo-go-samples/book-flight-ai-agent/proto"
 )
 
 var cfgEnv = conf.GetEnvironment()
