+++ book-flight-ai-agent/go-server/agents/cot_agent.go
@@ -26,12 +26,12 @@ import (
 )
 
 import (
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/actions"
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/conf"
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/model"
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/model/ollama"
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/prompts"
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/tools"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/actions"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/conf"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/model"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/model/ollama"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/prompts"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/tools"
 )
 
 type CotAgentRunner struct {
