+++ book-flight-ai-agent/go-server/model/ollama/options.go
@@ -17,7 +17,7 @@
 package ollama
 
 import (
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/model"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/model"
 )
 
 func WithStreamingFunc(fn model.CallFunc) model.Option {
