+++ book-flight-ai-agent/go-server/tools/bookingflight/booking_tools.go
@@ -24,7 +24,7 @@ import (
 )
 
 import (
-	"github.com/apache/dubbo-go-samples/llm/book-flight/go-server/tools"
+	"github.com/apache/dubbo-go-samples/book-flight-ai-agent/go-server/tools"
 )
 
 var (
